[テスト]実施途中

+++ app/controllers/public/orders_controller.rb
@@ -6,19 +6,24 @@ class Public::OrdersController < ApplicationController
 
   def create
     @order=Order.new(order_params)
+    @customer=current_customer
     @order.customer_id = current_customer.id
     #binding.pry
-    @order.save
-    current_customer.cart_items.each do |cart_item|
-      order_detail=OrderDetail.new
-      order_detail.item_id=cart_item.item_id
-      order_detail.order_id=@order.id
-      order_detail.quantity=cart_item.amount
-      order_detail.price=cart_item.item.taxprice
-      order_detail.save
+    if @order.save
+      current_customer.cart_items.each do |cart_item|
+        order_detail=OrderDetail.new
+        order_detail.item_id=cart_item.item_id
+        order_detail.order_id=@order.id
+        order_detail.quantity=cart_item.amount
+        order_detail.price=cart_item.item.taxprice
+        order_detail.save
+      end
+      current_customer.cart_items.destroy_all
+      redirect_to thanks_path
+    else
+      render :new
     end
-    current_customer.cart_items.destroy_all
-    redirect_to thanks_path
+
   end
 
   def confirm
